requirements.txt
Django>=5.0
requests>=2.32
pillow>=10.4.0